update main.js & move stories to stories/Divider
