+++ packages/react-components/react-divider/src/stories/Divider/DividerCustomStyles.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { shorthands, makeStyles } from '@griffel/react';
 import { tokens } from '@fluentui/react-theme';
-import { Divider } from '../Divider';
+import { Divider } from '../../Divider';
 
 const useStyles = makeStyles({
   root: {
