+++ packages/react-components/react-divider/src/stories/Divider/index.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta } from '@storybook/react';
-import { Divider } from '../Divider';
+import { Divider } from '../../Divider';
 import descriptionMd from './DividerDescription.md';
 export { Default } from './DividerDefault.stories';
 export { Vertical } from './DividerVertical.stories';
